implement show more strategy in show more view model
